Relationships and conditions with caching
- cached? now checks if there maybe is a relationship path cached too
- Added conditions, just add :conditions => bla, it will become a query
  string
- Conditions are cached and invalidated after save/destroy, found by
  cross referencing
- Added multiple helper methods to relationships. This needs some
  refactoring! Is not nice code.
- Refactored Dada instance cache invalidation methods to single method
- Adding a child to a parent now properly sets a relationship
- Also added helpers for getting all belongs_to relationships
- Made request handling a bit nicer, added named args to so some

+++ lib/dada/rest_helpers.rb
@@ -65,14 +65,15 @@ module Dada
 
       # wraps rest requests to the corresponding service
       # *emerging*
-      def handle_request(method, path, data = nil)
-        if(method == :get)
-          response = rest_client.get(path)
-        elsif(method == :post)
-          response = rest_client.post(path, :body => data)
-        elsif(method == :put)
-          response = rest_client.put(path, :body => data)
-        elsif(method == :delete)
+      def handle_request(method, path, opts={})
+        case method
+        when :get
+          response = rest_client.get(path, opts)
+        when :post
+          response = rest_client.post(path, opts)
+        when :put
+          response = rest_client.put(path, opts)
+        when :delete
           response = rest_client.delete(path)
         else
           raise UnknownRESTMethod, "handle_request only responds to get, put, post and delete"
